fix single fetch
