+++ src/main/scala/rift2Core/frontend/IF2.scala
@@ -25,18 +25,47 @@ import chisel3.util._
 
 import rift2Core.define._
 import rift2Core.L1Cache._
-
+import rift2Core.privilege._
 
 import chipsalliance.rocketchip.config.Parameters
 import freechips.rocketchip.diplomacy._
 import freechips.rocketchip.tilelink._
 
+abstract class IF2Base(edge: TLEdgeOut)(implicit p: Parameters) extends IcacheModule {
+  val iEdge = edge
+
+  val io = IO(new Bundle {
+
+    val if2_req  = Flipped(new DecoupledIO( new IF1_Bundle ))
+    val if2_resp = Vec(4, new DecoupledIO(new IF2_Bundle) )
+
+    val if_mmu = DecoupledIO(new Info_mmu_req)
+    val mmu_if = Flipped(DecoupledIO(new Info_mmu_rsp))
+
+    val if_cmm = Output(new Info_if_cmm)
 
 
-class IF2(edge: TLEdgeOut)(implicit p: Parameters) extends Icache(edge = edge){
+    val icache_get    = new DecoupledIO(new TLBundleA(iEdge.bundle))
+    val icache_access = Flipped(new DecoupledIO(new TLBundleD(iEdge.bundle)))
 
+    val flush = Input(Bool())
+
+    val ifence = Input(Bool())
+
+    /** prefetch is not guarantee to be accepted by cache*/
+    val preFetch = ValidIO( new PreFetch_Req_Bundle )
+  })
+
+}
+
+trait IF2_PreFetch { this: IF2Base => 
   io.preFetch.valid := io.icache_get.fire
   io.preFetch.bits.paddr := io.icache_get.bits.address + "b100000".U
+}
+
+class IF2(edge: TLEdgeOut)(implicit p: Parameters) extends IF2Base(edge) with ICache with IF2_PreFetch {
+
+
 }
